pyproject.toml
[project]
name = "pywebview-msg-app"
version = "0.1.0"
description = "dependencies"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pywebview",
    "uvicorn[standard]",
    "fastapi",
    "psycopg2-binary",
    "sqlalchemy",
    "python-dotenv"
]
